Simplify PHPUnit Polyfill abstraction.
